+++ lib/rest/messaging/v1/service/usAppToPerson.d.ts
@@ -165,7 +165,7 @@ interface UsAppToPersonListInstanceCreateOptions {
   hasEmbeddedPhone: boolean;
   helpKeywords?: string | string[];
   helpMessage?: string;
-  messageFlow?: string;
+  messageFlow: string;
   messageSamples: string | string[];
   optInKeywords?: string | string[];
   optInMessage?: string;