+++ spec/integration/rest/taskrouter/v1/workspace/worker.spec.js
@@ -54,6 +54,7 @@ describe('Worker', function() {
                   'date_created': '2017-05-30T23:05:29Z',
                   'date_updated': '2017-05-30T23:05:29Z',
                   'date_status_changed': '2017-05-30T23:05:29Z',
+                  'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',
@@ -99,6 +100,7 @@ describe('Worker', function() {
                   'date_created': '2017-05-30T23:05:29Z',
                   'date_updated': '2017-05-30T23:05:29Z',
                   'date_status_changed': '2017-05-30T23:05:29Z',
+                  'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',
@@ -149,6 +151,7 @@ describe('Worker', function() {
                   'date_created': '2017-05-30T23:05:29Z',
                   'date_updated': '2017-05-30T23:05:29Z',
                   'date_status_changed': '2017-05-30T23:05:29Z',
+                  'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',
@@ -216,6 +219,7 @@ describe('Worker', function() {
                   'date_created': '2017-05-30T23:05:29Z',
                   'date_updated': '2017-05-30T23:05:29Z',
                   'date_status_changed': '2017-05-30T23:05:29Z',
+                  'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
                   'links': {
                       'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',
@@ -310,6 +314,7 @@ describe('Worker', function() {
           'date_created': '2017-05-30T23:19:38Z',
           'date_updated': '2017-05-30T23:19:38Z',
           'date_status_changed': '2017-05-30T23:19:38Z',
+          'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',
@@ -373,6 +378,7 @@ describe('Worker', function() {
           'date_updated': '2017-05-30T23:32:39Z',
           'friendly_name': 'NewWorker3',
           'sid': 'WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
+          'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'workspace_sid': 'WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
@@ -440,6 +446,7 @@ describe('Worker', function() {
           'date_created': '2017-05-30T23:32:22Z',
           'date_updated': '2017-05-31T00:05:57Z',
           'date_status_changed': '2017-05-30T23:32:22Z',
+          'operating_unit_sid': 'OUaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'url': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa',
           'links': {
               'channels': 'https://taskrouter.twilio.com/v1/Workspaces/WSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Workers/WKaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels',